Create new exceptions for connect, fix error messages
