+++ src/aiosmtplib/errors.py
@@ -15,6 +15,7 @@ __all__ = (
     "SMTPSenderRefused",
     "SMTPServerDisconnected",
     "SMTPTimeoutError",
+    "SMTPConnectTimeoutError",
 )
 
 
@@ -50,6 +51,20 @@ class SMTPTimeoutError(SMTPException, TimeoutError):
     A timeout occurred while performing a network operation.
     """
 
+    pass
+
+
+class SMTPConnectTimeoutError(SMTPTimeoutError, SMTPConnectError):
+    """
+    A timeout occurred while connection to the SMTP server
+    """
+    pass
+
+
+class SMTPReadTimeoutError(SMTPTimeoutError):
+    """
+    A timeout occurred while wait SMTP server ready message
+    """
     pass
